+++ src/frontend/src/common/components/library/data_subview/components/EditTreeConfirmationModal/index.tsx
@@ -2,8 +2,8 @@ import CloseIcon from "@material-ui/icons/Close";
 import { Button } from "czifui";
 import React, { useEffect, useState } from "react";
 import { useEditTree } from "src/common/queries/trees";
-import BaseDialog from "src/components/BaseActionDialog";
-import { StyledIconButton } from "src/components/BaseActionDialog/style";
+import BaseDialog from "src/components/BaseDialog";
+import { StyledIconButton } from "src/components/BaseDialog/style";
 import Notification from "src/components/Notification";
 import { TreeNameInput } from "src/components/TreeNameInput";
 import { StyledTitle } from "./style";
