+++ src/frontend/src/components/BaseDialog/style.ts
